books/requirements.txt
asgiref==3.8.1
coverage==7.6.8
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.2
